requirements.txt
asgiref==3.8.1
attrs==24.2.0
cattrs==24.1.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==5.1.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
flatbuffers==24.3.25
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
numpy==2.1.1
openmeteo_requests==1.3.0
openmeteo_sdk==1.15.0
packaging==24.1
pandas==2.2.2
platformdirs==4.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
ratelim==0.1.6
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.3
